Separate the datasources logic from chat.py into commandsa and datasource utils
