+++ tests/test_datasource.py
@@ -11,22 +11,24 @@ from rich.console import Console
 
 from anton.chat import (
     ChatSession,
-    _DS_KNOWN_VARS,
-    _DS_SECRET_VARS,
-    _PROMPT_RECONNECT_CANCEL,
-    _build_datasource_context,
-    _handle_add_custom_datasource,
-    _handle_connect_datasource,
-    _handle_test_datasource,
-    _register_secret_vars,
-    _restore_namespaced_env,
-    _run_connection_test,
-    _scrub_credentials,
-    parse_connection_slug,
 )
 from anton.commands.datasource import (
+    _PROMPT_RECONNECT_CANCEL,
+    handle_add_custom_datasource as _handle_add_custom_datasource,
+    handle_connect_datasource as _handle_connect_datasource,
+    handle_test_datasource as _handle_test_datasource,
+    run_connection_test as _run_connection_test,
     handle_list_data_sources,
     handle_remove_data_source,
+)
+from anton.datasource_utils import (
+    _DS_KNOWN_VARS,
+    _DS_SECRET_VARS,
+    _build_datasource_context,
+    _register_secret_vars,
+    _restore_namespaced_env,
+    _scrub_credentials,
+    parse_connection_slug,
 )
 from anton.cli import app as _cli_app
 from anton.data_vault import DataVault, _slug_env_prefix
@@ -571,9 +573,9 @@ class TestHandleConnectDatasource:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=DataVault(vault_dir=vault_dir)),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", return_value="MySQL"),
+            patch("anton.commands.datasource.DataVault", return_value=DataVault(vault_dir=vault_dir)),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", return_value="MySQL"),
         ):
             result = await _handle_connect_datasource(
                 console, session._scratchpads, session
@@ -591,10 +593,10 @@ class TestHandleConnectDatasource:
         responses = iter(["PostgreSQL", "n", "n", "db.example.com", "", "", "", "", ""])
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -636,10 +638,10 @@ class TestHandleConnectDatasource:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -690,10 +692,10 @@ class TestHandleConnectDatasource:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -733,10 +735,10 @@ class TestHandleConnectDatasource:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -774,10 +776,10 @@ class TestHandleConnectDatasource:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -801,10 +803,10 @@ class TestHandleConnectDatasource:
         responses = iter(["HubSpot", "1", "n", "y", "pat-na1-abc123"])
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -843,10 +845,10 @@ class TestHandleConnectDatasource:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -935,10 +937,10 @@ class TestCredentialScrubbing:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -999,8 +1001,8 @@ class TestActiveDatasourceScoping:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry"),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry"),
         ):
             result = await _handle_connect_datasource(
                 console, session._scratchpads, session, prefill="hubspot-2"
@@ -1030,8 +1032,8 @@ class TestActiveDatasourceScoping:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry"),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry"),
         ):
             result = await _handle_connect_datasource(
                 console, session._scratchpads, session, prefill="hubspot-2"
@@ -1049,7 +1051,7 @@ class TestActiveDatasourceScoping:
         vault.save("oracle", "1", {"host": "oracle.host"})
         vault.save("hubspot", "2", {"access_token": "pat-xxx"})
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context()
 
         assert "oracle-1" in ctx
@@ -1061,7 +1063,7 @@ class TestActiveDatasourceScoping:
         vault.save("oracle", "1", {"host": "oracle.host"})
         vault.save("hubspot", "2", {"access_token": "pat-xxx"})
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context(active_only="hubspot-2")
 
         assert "hubspot-2" in ctx
@@ -1072,7 +1074,7 @@ class TestActiveDatasourceScoping:
         vault = DataVault(vault_dir=vault_dir)
         vault.save("oracle", "1", {"host": "oracle.host"})
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context(active_only="hubspot-99")
 
         assert "oracle-1" not in ctx
@@ -1083,7 +1085,7 @@ class TestActiveDatasourceScoping:
             "postgres", "prod_db", {"host": "pg.example.com", "password": "s3cr3t"}
         )
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context()
 
         assert "DS_POSTGRES_PROD_DB__HOST" in ctx
@@ -1094,7 +1096,7 @@ class TestActiveDatasourceScoping:
         vault = DataVault(vault_dir=vault_dir)
         vault.save("postgres", "prod_db", {"host": "pg.example.com"})
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context()
 
         assert "postgres-prod_db" in ctx
@@ -1106,7 +1108,7 @@ class TestActiveDatasourceScoping:
         vault.save("postgres", "prod_db", {"host": "pg.example.com"})
         vault.save("hubspot", "main", {"access_token": "pat-abc"})
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             ctx = _build_datasource_context()
 
         assert "postgres-prod_db" in ctx
@@ -1134,7 +1136,7 @@ class TestCliCommandRegistration:
 class TestHandleListDataSources:
     def test_empty_vault_shows_message(self, vault_dir):
         console = MagicMock()
-        with patch("anton.chat.DataVault", return_value=DataVault(vault_dir=vault_dir)):
+        with patch("anton.commands.datasource.DataVault", return_value=DataVault(vault_dir=vault_dir)):
             handle_list_data_sources(console)
         printed = " ".join(str(c) for c in console.print.call_args_list)
         assert "No data sources" in printed or "connect" in printed
@@ -1208,8 +1210,8 @@ class TestHandleTestDatasource:
         scratchpads.get_or_create = AsyncMock(return_value=pad)
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(console, scratchpads, "postgresql-prod_db")
 
@@ -1236,8 +1238,8 @@ class TestHandleTestDatasource:
         scratchpads.get_or_create = AsyncMock(return_value=pad)
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(console, scratchpads, "postgresql-prod_db")
 
@@ -1251,8 +1253,8 @@ class TestHandleTestDatasource:
         scratchpads = AsyncMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(console, scratchpads, "postgresql-ghost")
 
@@ -1265,8 +1267,8 @@ class TestHandleTestDatasource:
         scratchpads = AsyncMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=DataVault(vault_dir=vault_dir)),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=DataVault(vault_dir=vault_dir)),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(console, scratchpads, "")
 
@@ -1310,10 +1312,10 @@ class TestEditDatasourceFlow:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(prompt_values)),
             ),
         ):
@@ -1365,10 +1367,10 @@ class TestEditDatasourceFlow:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(prompt_values)),
             ),
         ):
@@ -1393,8 +1395,8 @@ class TestEditDatasourceFlow:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             result = await _handle_connect_datasource(
                 console,
@@ -1492,10 +1494,10 @@ class TestEnvActivationCollisionFree:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -1537,8 +1539,8 @@ class TestEnvActivationCollisionFree:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_connect_datasource(
                 console,
@@ -1560,8 +1562,8 @@ class TestDatasourceSlashCommandBehavior:
         scratchpads = AsyncMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=DataVault(vault_dir=vault_dir)),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=DataVault(vault_dir=vault_dir)),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(console, scratchpads, "")
 
@@ -1577,9 +1579,9 @@ class TestDatasourceSlashCommandBehavior:
         console = MagicMock()
 
         with (
-            patch("anton.chat.DataVault", return_value=DataVault(vault_dir=vault_dir)),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", return_value="UnknownEngine"),
+            patch("anton.commands.datasource.DataVault", return_value=DataVault(vault_dir=vault_dir)),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", return_value="UnknownEngine"),
         ):
             updated = await _handle_connect_datasource(
                 console,
@@ -1686,7 +1688,7 @@ class TestTemporaryFlatExecution:
         assert os.environ.get("DS_HOST") == "analytics.example.com"
         assert "DS_POSTGRES_ANALYTICS__HOST" not in os.environ
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             _restore_namespaced_env(vault)
 
         assert "DS_HOST" not in os.environ
@@ -1700,7 +1702,7 @@ class TestTemporaryFlatExecution:
 
         vault.inject_env("postgres", "prod_db", flat=True)
 
-        with patch("anton.chat.DataVault", return_value=vault):
+        with patch("anton.datasource_utils.DataVault", return_value=vault):
             _restore_namespaced_env(vault)
 
         assert "DS_HOST" not in os.environ
@@ -1744,8 +1746,8 @@ class TestTemporaryFlatExecution:
         scratchpads.get_or_create = AsyncMock(return_value=pad)
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
         ):
             await _handle_test_datasource(
                 MagicMock(), scratchpads, "postgresql-prod_db"
@@ -1780,7 +1782,7 @@ class TestStaleDsRegistrationState:
         )
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1790,7 +1792,7 @@ class TestStaleDsRegistrationState:
         vault.delete("postgresql", "prod_db")
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1813,7 +1815,7 @@ class TestStaleDsRegistrationState:
         )
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1835,7 +1837,7 @@ class TestStaleDsRegistrationState:
         )
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1859,7 +1861,7 @@ class TestStaleDsRegistrationState:
         )
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1867,7 +1869,7 @@ class TestStaleDsRegistrationState:
         known_after_first = len(_DS_KNOWN_VARS)
 
         with patch(
-            "anton.datasource_registry.DatasourceRegistry", return_value=registry
+            "anton.datasource_utils.DatasourceRegistry", return_value=registry
         ):
             _restore_namespaced_env(vault)
 
@@ -1930,10 +1932,10 @@ class TestAddCustomDatasourceFlow:
 
         with (
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_add_custom_datasource(
@@ -1970,10 +1972,10 @@ class TestAddCustomDatasourceFlow:
 
         with (
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_add_custom_datasource(
@@ -2015,10 +2017,10 @@ class TestAddCustomDatasourceFlow:
 
         with (
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_add_custom_datasource(
@@ -2103,16 +2105,16 @@ class TestCustomDatasourceConnectFlow:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
             patch(
-                "anton.chat.DatasourceRegistry",
+                "anton.commands.datasource.DatasourceRegistry",
                 return_value=self._make_registry(tmp_path),
             ),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_connect_datasource(
@@ -2159,16 +2161,16 @@ class TestCustomDatasourceConnectFlow:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
             patch(
-                "anton.chat.DatasourceRegistry",
+                "anton.commands.datasource.DatasourceRegistry",
                 return_value=self._make_registry(tmp_path),
             ),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_connect_datasource(
@@ -2220,16 +2222,16 @@ class TestCustomDatasourceConnectFlow:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
             patch(
-                "anton.chat.DatasourceRegistry",
+                "anton.commands.datasource.DatasourceRegistry",
                 return_value=self._make_registry(tmp_path),
             ),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             result = await _handle_connect_datasource(
@@ -2272,16 +2274,16 @@ class TestCustomDatasourceConnectFlow:
         responses = iter(["0", "My API Service", "I have an API key", "n", "my_key"])
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
             patch(
-                "anton.chat.DatasourceRegistry",
+                "anton.commands.datasource.DatasourceRegistry",
                 return_value=self._make_registry(tmp_path),
             ),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
-            patch("anton.chat.Path") as mock_path_cls,
+            patch("anton.commands.datasource.Path") as mock_path_cls,
         ):
             self._mock_ds_path(mock_path_cls, tmp_path)
             await _handle_connect_datasource(console, session._scratchpads, session)
@@ -2323,10 +2325,10 @@ class TestEditDatasourceWithTestSnippet:
         )  # field values, then retry?
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(responses)),
             ),
         ):
@@ -2367,10 +2369,10 @@ class TestEditDatasourceWithTestSnippet:
         )
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(prompt_responses)),
             ),
         ):
@@ -2419,9 +2421,9 @@ class TestEditDatasourceWithTestSnippet:
         }
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", new=AsyncMock(return_value="n")),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", new=AsyncMock(return_value="n")),
         ):
             result = await _run_connection_test(
                 console,
@@ -2450,9 +2452,9 @@ class TestPromptCopyConsistency:
         vault = DataVault(vault_dir=vault_dir)
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", new=AsyncMock(return_value=None)),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", new=AsyncMock(return_value=None)),
         ):
             result = await _handle_connect_datasource(
                 console, session._scratchpads, session
@@ -2474,9 +2476,9 @@ class TestPromptCopyConsistency:
         session._scratchpads.get_or_create = AsyncMock(return_value=pad)
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", new=AsyncMock(return_value=None)),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", new=AsyncMock(return_value=None)),
         ):
             engine_def = registry.get("postgresql")
             credentials = {
@@ -2510,10 +2512,10 @@ class TestPromptCopyConsistency:
         poc_calls = iter(["PostgreSQL", None])  # engine selected, then Esc
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
             patch(
-                "anton.chat.prompt_or_cancel",
+                "anton.commands.datasource.prompt_or_cancel",
                 new=AsyncMock(side_effect=lambda *a, **kw: next(poc_calls)),
             ),
         ):
@@ -2540,9 +2542,9 @@ class TestPromptCopyConsistency:
             return None  # Esc on every prompt to bail out
 
         with (
-            patch("anton.chat.DataVault", return_value=vault),
-            patch("anton.chat.DatasourceRegistry", return_value=registry),
-            patch("anton.chat.prompt_or_cancel", new=AsyncMock(side_effect=_capture)),
+            patch("anton.commands.datasource.DataVault", return_value=vault),
+            patch("anton.commands.datasource.DatasourceRegistry", return_value=registry),
+            patch("anton.commands.datasource.prompt_or_cancel", new=AsyncMock(side_effect=_capture)),
         ):
             # "PostgreeSQL" triggers fuzzy match against "PostgreSQL"
             await _handle_connect_datasource(
